Добавляет функц-ный компонент MoviesList и рендерит в него карточки фильмов
