+++ src/components/main/main.jsx
@@ -1,19 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import MoviesList from '../movies-list/movies-list.jsx';
 import Card from '../card/card.jsx';
 
 const Main = ({films, promoCard, mainTitleClickHandler}) => {
   const {title, genre, releaseYear} = promoCard;
-
-  const cardsList = films.map((film) => (
-    <Card
-      title={film.title}
-      key={Math.ceil(Math.random() * 1000)}
-      poster={film.preview}
-      mainTitleClickHandler={mainTitleClickHandler}
-    />
-  ));
 
   return (
     <>
@@ -100,9 +92,12 @@ const Main = ({films, promoCard, mainTitleClickHandler}) => {
               <a href="#" className="catalog__genres-link">Thrillers</a>
             </li>
           </ul>
-          <div className="catalog__movies-list">
-            {cardsList}
-          </div>
+
+          <MoviesList
+            films={films}
+            mainTitleClickHandler={mainTitleClickHandler}
+          />
+
           <div className="catalog__more">
             <button className="catalog__button" type="button">Show more</button>
           </div>
